Befor After Module Extra Images

+++ site_files/app/Helpers/ImageUploader.php
@@ -35,7 +35,9 @@ class ImageUploader
         })->save($destinationPath . '/' . $webpFileName);
         /**************************** */
         /**************************** */
-        File::delete($destinationPath . '/' . $fileName);
+        if($destinationPath . '/' . $webpFileName !== $destinationPath . '/' . $fileName){
+            File::delete($destinationPath . '/' . $fileName);
+        }
         /**************************** */
         /**************************** */
         if ($makeOtherSizesImages === true) {
